% Generated by build script
\name{osf_nodes}
\alias{osf_nodes}
\title{Search OSF nodes (projects and components)}
\usage{
osf_nodes()
}
\arguments{
\item{query}{Title search filter}
\item{page_size}{Number of results per page (default 25, max 100)}
}
\value{tibble: id, title, description, category, date_created,}
\description{Search OSF nodes (projects and components)}
